+++ paper.bib
@@ -19,7 +19,8 @@
   volume={228},
   pages={111099},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.ijsolstr.2021.111099}
 }
 
 @article{bai2020chemo,
@@ -29,7 +30,8 @@
   volume={183},
   pages={45--49},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.scriptamat.2020.03.027}
 }
 
 @article{miehe2010phase,
@@ -40,7 +42,8 @@
   number={45-48},
   pages={2765--2778},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.cma.2010.04.011}
 }
 
 @article{kobayashi1993modeling,
@@ -51,5 +54,6 @@
   number={3-4},
   pages={410--423},
   year={1993},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/0167-2789(93)90120-P}
 }
