### 2022-02-17 #4175 @yangbai90
The cloc output for pure AsFem code (without external packages) is shown below:

```
Language                     files          blank        comment           code
GLSL                            27             80            391          21663
C++                            206           2324           5687          21116
SWIG                            97            958            207          10384
C/C++ Header                   176           1649           6366           5887
CMake                            2             66            272            497
Python                           9             81            108            496
YAML                             1              4             10            200
Markdown                         2             42              0            104
TeX                              1              4              0             55
SUM:                           521           5208          13041          60402
```
