### 2022-02-28 #4175 @Kevin-Mattheus-Moerman
@likask, @tjhei, @chrisrichardson, @vijaysm, @diehlpk would you be interested in reviewing this work for JOSS?
