### 2022-03-01 #4175 @xuanxu
> is @editorialbot responding to deleted comments here? @Kevin-Mattheus-Moerman has been having some weird behaviour issues.

Maybe, the bot receives the comments immediately after they are posted, so even if the comment is deleted later @editorialbot will respond.
