### 2022-03-01 #4175 @Kevin-Mattheus-Moerman
@xuanxu yes it is repeatable see :point_up: i.e. the error comes up again when one deletes the comments causing it and the error message itself (not sure which is responsible).
