### 2022-03-18 #4175 @Kevin-Mattheus-Moerman
@schruste, @mikaem, @likask, @tjhei, @diehlpk, would you be interested in reviewing this work featuring a C++ library for finite element analysis and multiphysics for [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/)? The [JOSS review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#reviewing-for-joss) focuses on evaluation of [the software](https://github.com/M3Group/AsFem) as well as a [short paper](https://github.com/openjournals/joss-papers/blob/joss.04175/joss.04175/10.21105.joss.04175.pdf).

Note we can be flexible regarding time needed for review.

JOSS reviews are usually smooth and streamlined and take place on GitHub. Here is an example of a completed review: https://github.com/openjournals/joss-reviews/issues/1441
