+++ include/Utils/RankFourTensor.h
@@ -26,14 +26,26 @@
 
 using namespace std;
 
+/**
+ * forward declaration
+ */
 class RankTwoTensor;
 
+/**
+ * this class implement the general operators/calculations in Rank-4 tensor
+ */
 class RankFourTensor{
 public:
+    /**
+     * constructor for different purpose
+     */
     RankFourTensor();
     RankFourTensor(const double &val);
     RankFourTensor(const RankFourTensor &a);
     
+    /**
+     * enum for different initializing
+     */
     enum InitMethod{
         InitZero,
         InitIdentity,
@@ -43,6 +55,13 @@ public:
     };
     RankFourTensor(const InitMethod &method);
 
+    /**
+     * get i-k components for jacobian matrix from a rank-4 tensor
+     * @param i integer for row-index
+     * @param j integer for col-index
+     * @param grad_test gradient of the test function
+     * @param grad_trial gradient of the tril function
+     */
     inline double GetIKjlComponent(const int &i,const int &k,
                                    const Vector3d &grad_test,
                                    const Vector3d &grad_trial)const{
@@ -67,6 +86,13 @@ public:
     //*** For operator overload
     //********************************************
     // for index based access
+    /**
+     * () operator for rank-4 tensor \f$\mathbb{C}_{ijkl}\f$
+     * @param i i index
+     * @param j j index
+     * @param k k index
+     * @param l l index
+     */
     inline double operator()(const int &i,const int &j,const int &k,const int &l) const{
         if(i<1||i>3 || j<1||j>3 || k<1||k>3 || l<1||l>3){
             MessagePrinter::PrintErrorTxt("your i or j or k or l is out of range when you call a rank-4 tensor");
@@ -74,6 +100,13 @@ public:
         }
         return _vals[(((i-1)*_N+j-1)*_N+k-1)*_N+l-1];
     }
+    /**
+     * () operator for rank-4 tensor \f$\mathbb{C}_{ijkl}\f$
+     * @param i i index
+     * @param j j index
+     * @param k k index
+     * @param l l index
+     */
     inline double& operator()(const int &i,const int &j,const int &k,const int &l){
         if(i<1||i>3 || j<1||j>3 || k<1||k>3 || l<1||l>3){
             MessagePrinter::PrintErrorTxt("your i or j or k or l is out of range when you call a rank-4 tensor");
@@ -81,11 +114,24 @@ public:
         }
         return _vals[(((i-1)*_N+j-1)*_N+k-1)*_N+l-1];
     }
+    /**
+     * get the voigt component of a rank-4 tensor \f$\mathbb{C}_{ijkl}\f$
+     * @param i i index
+     * @param j j index
+     */
     double VoigtIJcomponent(const int &i,const int &j) const;
     // for components based access
+    /**
+     * [] operator to get the component of a rank-4 tensor \f$\mathbb{C}_{ijkl}\f$
+     * @param i i index
+     */
     inline double  operator[](const int &i) const{
         return _vals[i-1];
     }
+    /**
+     * [] operator to get the component of a rank-4 tensor \f$\mathbb{C}_{ijkl}\f$
+     * @param i i index
+     */
     inline double& operator[](const int &i){
         return _vals[i-1];
     }
@@ -94,72 +140,141 @@ public:
         return _vals[i];
     }
     //*** for =
+    /**
+     * = operator to a rank-4 tensor \f$\mathbb{C}_{ijkl}\f$
+     * @param a right hand side scalar
+     */
     inline RankFourTensor& operator=(const double &a){
[…]
