+++ src/FE/QPoints/CreateGaussLegendrePoint.cpp
@@ -15,9 +15,8 @@
 
 #include "FE/QPointGaussLegendre.h"
 
-void QPointGaussLegendre::CreateQpoints(MeshType meshtype)
-{
-    if (!_HasDim && !_HasOrder && !_HasSettings){
+void QPointGaussLegendre::CreateQpoints(MeshType meshtype){
+    if (!_HasDim || !_HasOrder || !_HasSettings){
         MessagePrinter::PrintErrorTxt("can\'t generate gauss-legendre points for the integration, neither the dim, the order or the type is given");
         MessagePrinter::AsFem_Exit();
     }
@@ -34,15 +33,15 @@ void QPointGaussLegendre::CreateQpoints(MeshType meshtype)
 //****************************************************************
 //*** For 1D Gauss-Legendre point generation
 //****************************************************************
-void QPointGaussLegendre::Create1DGaussPoint()
-{
+void QPointGaussLegendre::Create1DGaussPoint(){
+    // taken from https://en.wikipedia.org/wiki/Gauss%E2%80%93Legendre_quadrature
     switch (GetQpOrder())
     {
     case 0:
     case 1:
     {
         _nQpPoints = 1;
-        _QpCoords.resize(2, 0.0);
+        _QpCoords.resize(1*(1+1), 0.0);
         (*this)(1, 0) = 2.0;
         (*this)(1, 1) = 0.0;
         return;
@@ -51,9 +50,10 @@ void QPointGaussLegendre::Create1DGaussPoint()
     case 3:
     {
         _nQpPoints = 2;
-        _QpCoords.resize(2 * (1 + 1), 0.0);
+        _QpCoords.resize(2*(1+1), 0.0);
         (*this)(1, 0) = 1.0;
-        (*this)(1, 1) = -sqrt(1.0 / 3.0);
+        (*this)(1, 1) =-sqrt(1.0 / 3.0);
+
         (*this)(2, 0) = 1.0;
         (*this)(2, 1) = sqrt(1.0 / 3.0);
         return;
@@ -62,13 +62,14 @@ void QPointGaussLegendre::Create1DGaussPoint()
     case 5:
     {
         _nQpPoints = 3;
-        _QpCoords.resize(3 * (1 + 1), 0.0);
-        (*this)(1, 0) = 5.0 / 9.0;
-        (*this)(1, 1) = -sqrt(0.6);
-        (*this)(2, 0) = 8.0 / 9.0;
+        _QpCoords.resize(3*(1+1), 0.0);
+        (*this)(1, 0) = 5.0/9.0;
+        (*this)(1, 1) =-sqrt(0.6);
+
+        (*this)(2, 0) = 8.0/9.0;
         (*this)(2, 1) = 0.0;
 
-        (*this)(3, 0) = 5.0 / 9.0;
+        (*this)(3, 0) = 5.0/9.0;
         (*this)(3, 1) = sqrt(0.6);
         return;
     }
@@ -76,60 +77,113 @@ void QPointGaussLegendre::Create1DGaussPoint()
     case 7:
     {
         _nQpPoints = 4;
-        _QpCoords.resize(4 * (GetDim() + 1), 0.0);
-        const static double t = sqrt(4.8);
+        _QpCoords.resize(4*(1+1),0.0);
+        const double t1=sqrt(3.0/7.0-(2.0/7.0)*sqrt(6.0/5.0));
+        const double t2=sqrt(3.0/7.0+(2.0/7.0)*sqrt(6.0/5.0));
+        const double w1=(18.0+sqrt(30.0))/36.0;
+        const double w2=(18.0-sqrt(30.0))/36.0;
 
-        (*this)(1, 1) = -sqrt((3.0 + t) / 7.0);
[…]
