### 2022-04-06 aa1ba7e
add hand-write newton-raphson as an alternative to SNES
