+++ src/NonlinearSolver/NewtowRaphsonSolve.cpp
@@ -1,0 +1,155 @@
+//****************************************************************
+//* This file is part of the AsFem framework
+//* A Simple Finite Element Method program (AsFem)
+//* All rights reserved, Yang Bai/M3 Group @ CopyRight 2022
+//* https://github.com/M3Group/AsFem
+//* Licensed under GNU GPLv3, please see LICENSE for details
+//* https://www.gnu.org/licenses/gpl-3.0.en.html
+//****************************************************************
+//++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+//+++ Author : Yang Bai
+//+++ Date   : 2022.04.06
+//+++ Purpose: implement the hand-write newton-raphson method without
+//+++          any line search from PETSc
+//++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+#include "NonlinearSolver/NonlinearSolver.h"
+
+bool NonlinearSolver::NewtowRaphsonSolve(Mesh &mesh,DofHandler &dofHandler,
+            ElmtSystem &elmtSystem,MateSystem &mateSystem,
+            BCSystem &bcSystem,
+            SolutionSystem &solutionSystem,EquationSystem &equationSystem,
+            FE &fe,FESystem &feSystem,
+            FEControlInfo &fectrlinfo){
+    _monctx.iters=0;
+    bool IsConverge=false;
+    char buff[68];
+    string str;
+    while(_monctx.iters<_MaxIters && !IsConverge){
+        // preset new boundary condition to unew
+        bcSystem.ApplyPresetBC(mesh,dofHandler,FECalcType::ComputeResidual,
+                                fectrlinfo.t+fectrlinfo.dt,
+                                fectrlinfo.ctan,
+                                solutionSystem._Unew,
+                                equationSystem._AMATRIX,
+                                equationSystem._RHS);
+        // set the ctan array according to different time stepping method
+        if(fectrlinfo._timesteppingtype==TimeSteppingType::STATIC){
+            // for static analysis
+            fectrlinfo.ctan[0]=1.0;
+            fectrlinfo.ctan[1]=0.0;
+            fectrlinfo.ctan[2]=0.0;
+            VecCopy(solutionSystem._Unew,solutionSystem._Utemp);// in FormBulkFE, we always use Utemp for the calculation
+        }
+        else if(fectrlinfo._timesteppingtype==TimeSteppingType::BACKWARDEULER){
+            fectrlinfo.ctan[0]=1.0;
+            fectrlinfo.ctan[1]=1.0/fectrlinfo.dt;
+            fectrlinfo.ctan[2]=0.0;
+            // calculate the current velocity
+            VecWAXPY(solutionSystem._V,-1.0,solutionSystem._U,solutionSystem._Unew);//V=-Uold+Unew
+            VecScale(solutionSystem._V,fectrlinfo.ctan[1]);//V=V*1.0/dt
+            VecCopy(solutionSystem._Unew,solutionSystem._Utemp);// Used in FormBulkFE
+        }
+        else if(fectrlinfo._timesteppingtype==TimeSteppingType::CRANCKNICLSON){
+            fectrlinfo.ctan[0]=0.5;
+            fectrlinfo.ctan[1]=1.0/fectrlinfo.dt;
+            fectrlinfo.ctan[2]=0.0;
+            // calculate the current velocity
+            VecWAXPY(solutionSystem._V,-1.0,solutionSystem._U,solutionSystem._Unew);//V=-Uold+Unew
+            VecScale(solutionSystem._V,fectrlinfo.ctan[1]);//V=V*1.0/dt
+
[…]
