### 2022-05-10 #4252 @Kevin-Mattheus-Moerman
@mscroggs @vijaysm Thanks again for your help with this review. I see you have left several boxes unticked. Could you summarize what points the authors should work on? Or have you perhaps opened issues relating to these on the project repository? If so can you link to them here? Thanks.
