### 2022-05-16 M3Group/AsFem#62 @mscroggs
Provide a Docker image where AsFem can be used
Installing the requirements of AsFem (Petsc, etc) can take a while. It would nice for new users is there was a Docker image or some other similar package that could be downloaded to use the library without having to build everything from source
