### 2022-05-16 M3Group/AsFem#64 @mscroggs
State of the field in JOSS paper
There is not very much discussion of how AsFem compares to other finite element libraries in the JOSS paper. Libraries that you could mention include:

- FEniCS: [website](https://fenicsproject.org/), [GitHub repos](https://github.com/FEniCS), [paper](https://doi.org/10.11588/ans.2015.100.20553)
- MFEM: [website](https://mfem.org/), [GitHub repo](https://github.com/mfem/mfem/), [paper](https://doi.org/10.1016/j.camwa.2020.06.009)
- deal.ii: [website](https://www.dealii.org/), [GitHub repo](https://github.com/dealii/dealii), [paper](https://doi.org/10.1515/jnma-2021-0081)
- DUNE: [website](https://www.dune-project.org/), [GitHub repos](https://github.com/dune-project), [paper](https://doi.org/10.1016/j.camwa.2020.06.007)
- FreeFEM: [website](https://freefem.org/), [GitHub repos](https://github.com/FreeFem)
