### 2022-06-06 M3Group/AsFem#64 @vijaysm
I'd like to second this comment as a reviewer. There are a lot of very successful and tested, general-purpose FEM libraries out there. I fail to see a proper motivation in the JOSS paper on why there was a need to create a new one, instead of using the capabilities in the ones that @mscroggs has listed above. I'd probably also add [libMesh](https://github.com/libMesh/libmesh) and [MoFEM](http://mofem.eng.gla.ac.uk/mofem/html/) to that list as well given the wide variety of applications where they have been used.
