### 2022-06-06 M3Group/AsFem#62 @vijaysm
While installing PETSc is less of a hassle in most systems, having a docker image will allow interested users to get started quickly. So I agree that this should be made available even if building the dependencies is straightforward. You may be able to take a stock Ubuntu image and install a [version of PETSc](https://packages.ubuntu.com/search?keywords=PETSc) directly to build AsFEM.
