### 2022-06-20 M3Group/AsFem#62 @schruste
I agree with both, @mscroggs  and @vijaysm . Especially in view of a not too comfortable installation of AsFem and installation instructions that make several implicit assumptions (debian-based OS, etc..).
