### 2022-06-20 M3Group/AsFem#67 @schruste
AsFEM build infrastructure
The installation - even when considered together with the installation instructions -  is not very comfortable.

I managed to install AsFem after several attempts on my arch-based linux, but I did not like that I had to fix/correct several parts of the installation due to several implicit assumptions used that should be made explicit.
A few points:

* In the instructions a debian-based OS is assumed (see the prequisite packages and `apt install ...` instructions). I think restricting to debian-based systems is OK, but please add sentences that other OS-users will need to adapt...
* Several things are hardcoded in the cmake-build-files. I suggest to use the full power of cmake. cmake has variables that can be used to let cmake search for directories for dependecies (PetSC / MPI / ..). Global environment variables don't need to be used here. Currently global environment variables shall be used to described dependencies. I suggest to a) replace these with cmake-variables and b) to check for the existence of the correspondingly needed libraries in the corresponding location **before** starting the build process.
* A full setup script could also be provided, e.g. similar to the github-workflow-file.
* if a Docker-Setup is set up, see #62 I would also recommend to publish the full Docker-file as it is also a documentation of the required dependencies and the setup.
