### 2022-06-20 #4252 @schruste
@Kevin-Mattheus-Moerman : Sorry for being sooo late! I now found some time to continue my review. I added to two of @mscroggs issues on the AsFem-github and opened another issue.

My main points mostly coincide with @mscroggs assessment:
 * **Installation 1:** The installation is still quite painful. With some effort it is not too hard to setup, but the non-smoothness of the installation procedure can be an unnecessary barrier for interested users. There is room for improvement in the build system. CMake allows for more convenience than is used at the moment. Further, the documentation can be improved, see also the [corresponding issue](https://github.com/M3Group/AsFem/issues/67)
 * **Installation 2:** A Docker-image would be a (comparably simple) way to circumvent the setup overhead. At least for users that want to have a quick look and first try/run this would make perfect sense. For users that are convinced of the package the setup overhead may again be acceptable, but for new ones that simply want to try it out a docker-based version would be a good idea, see also [the issue opened by @mscroggs ](https://github.com/M3Group/AsFem/issues/62).
 * **Statement of need/field:** The author **has to explain** why this additional software package should be prefered in the considered use cases over the more obvious solutions (using established finite element frameworks with a lot of convenience (layers), like mfem, fenics / firedrake, ngsolve, freefem++, ... )´. Neither is this explained in the paper nor are there any proper references to similar libraries or techniques. See also [the issue opened by @mscroggs ](https://github.com/M3Group/AsFem/issues/64)
* There are some more details:
    * Is the code only low order and only continuous Lagrange elements?
    * The problem with the quadrature rule is only explained where it makes problems. I would recommend discussing the issue of quadrature in the documentation right away.
    * It seems to me that only a fixed number of pre-defined PDE-problems can be treated. Could this be (easily) extended to other PDE problems?
