### 2022-06-26 fcaf883
modify CMakeLists.txt for system installed mpi & petsc
