+++ docs/install.md
@@ -1,39 +1,48 @@
+# System requirement
+
+* Linux
+* Windows (only works for Cygwin+VisualStudio 2020 or newer)
+* MacOX (it should work, but we haven't tested it yet!)
+
+
 # Basic components
 
-Before we begin the installation, some components are needed, for example:
+Before we start the installation, some components are required, for example:
 
 * gcc g++ gfortran
 * cmake
 * git
 * python3
 
-<span style="color:red">In the following installation, please **do not** use `sudo` or `root`!!! It's only supposed to be used for `apt install`</span>.
+<span style="color:red">In the following installation, please **do not** use `sudo` or `root`!!! It's only supposed to be used for system installation like `sudo apt install`, `sudo zypper install`, `sudo yum install`, and so on</span>.
 
-**It is highly recommended to install openmpi and PETSc from the source code!**
+**It is highly recommended to install the openmpi and PETSc packages from the source code!**
 
 ## Install gcc
-You can use the steps below to install a modern GCC compiler if you don't already have one. **Otherwise, go straight to the MPI setup**.
+One can follow the steps list below to install a modern GCC compiler if one don't already have one. **Otherwise, please go straight to the MPI setup**.
 
 <span style="color:red">**Before you start, you'll need a GCC (the old one) compiler to compile another GCC, therefore if your system doesn't have one, please run(for Ubuntu)**</span>:
 ```
 sudo apt install gcc g++ gfortran cmake git python3 build-essential
 ```
-For ubuntu 20.04, `apt install` will bring you the `GCC9.3.0` package which is **new** and **good** enough for our installation, so you can skip the GCC installation part list below. Anyway, you can also install a newer version, i.e., GCC10.3.0 or even newer.
+For ubuntu 22.04, `apt install` will bring you the `GCC11.2.0` package which is **new** and **good** enough for our installation, so you can skip the GCC installation part list below. Anyway, you can also install a newer version, i.e., GCC12.1.0 or even newer.
 
-To begin, you should download and unzip the gcc source code:
+<span style="color:red">**The installation itself is not limited to Ubuntu system, for other linux distributions, one need to change the command to the consistent one. For instance, `sudo zypper`, `sudo yum`, `sudo dnf`...**</span>:
+
+To begin, one should download and unzip the gcc source code:
 ```
-curl -L -O http://mirrors.concertpass.com/gcc/releases/gcc-10.3.0/gcc-10.3.0.tar.gz
+curl -L -O http://mirrors.concertpass.com/gcc/releases/gcc-11.3.0/gcc-11.3.0.tar.gz
 
-tar -xf gcc-10.3.0.tar.gz
+tar -xf gcc-11.3.0.tar.gz
 ```
 or directly go to the official website [gcc](https://gcc.gnu.org/).
 
-Next, we'll need to get the GCC pre-request:
+Next, we'll need to download the GCC pre-request as follows:
 ```
-cd gcc-10.3.0
+cd gcc-11.3.0
 ./contrib/download_prerequisites
 ```
-we can then config, build and install GCC:
+we can then configure, build and install GCC:
 ```
 ./configure --prefix=**your-gcc-install-path** \
 --disable-multilib \
[…]
