### 2022-06-26 M3Group/AsFem#67 @yangbai90
Dear @schruste , thanks for the suggestion, I have updated the CMakeList file according to your comments.
