### 2022-06-26 2feb418
enable make test in CMakeList, after make, one could also run make test
