### 2022-10-19 0a55867
tremendous changes for v0.8, please dont clone this one for simulation, its still under construction
