+++ paper.bib
@@ -34,6 +34,17 @@
   doi={10.1016/j.scriptamat.2020.03.027}
 }
 
+@article{bai2019two,
+  title={Two-level modeling of lithium-ion batteries},
+  author={Bai, Yang and Zhao, Ying and Liu, Wei and Xu, Bai-Xiang},
+  journal={Journal of Power Sources},
+  volume={422},
+  pages={92--103},
+  year={2019},
+  publisher={Elsevier},
+  doi={10.1016/j.jpowsour.2019.03.026}
+}
+
 @article{miehe2010phase,
   title={A phase field model for rate-independent crack propagation: Robust algorithmic implementation based on operator splits},
   author={Miehe, Christian and Hofacker, Martina and Welschinger, Fabian},
@@ -56,4 +67,58 @@
   year={1993},
   publisher={Elsevier},
   doi={10.1016/0167-2789(93)90120-P}
+}
+
+@article{alnaes2015fenics,
+  title={The FEniCS project version 1.5},
+  author={Aln{\ae}s, Martin and Blechta, Jan and Hake, Johan and Johansson, August and Kehlet, Benjamin and Logg, Anders and Richardson, Chris and Ring, Johannes and Rognes, Marie E and Wells, Garth N},
+  journal={Archive of Numerical Software},
+  volume={3},
+  number={100},
+  year={2015},
+  doi={10.11588/ans.2015.100.20553}
+}
+@article{anderson2021mfem,
+  title={MFEM: A modular finite element methods library},
+  author={Anderson, Robert and Andrej, Julian and Barker, Andrew and Bramwell, Jamie and Camier, Jean-Sylvain and Cerveny, Jakub and Dobrev, Veselin and Dudouit, Yohann and Fisher, Aaron and Kolev, Tzanio and others},
+  journal={Computers \& Mathematics with Applications},
+  volume={81},
+  pages={42--74},
+  year={2021},
+  publisher={Elsevier},
+  doi={10.1016/j.camwa.2020.06.009}
+}
+@article{arndt2021deal,
+  title={The deal. II library, version 9.3},
+  author={Arndt, Daniel and Bangerth, Wolfgang and Blais, Bruno and Fehling, Marc and Gassm{\"o}ller, Rene and Heister, Timo and Heltai, Luca and K{\"o}cher, Uwe and Kronbichler, Martin and Maier, Matthias and others},
+  journal={Journal of Numerical Mathematics},
+  volume={29},
+  number={3},
+  pages={171--186},
+  year={2021},
+  publisher={De Gruyter},
+  doi={10.1515/jnma-2021-0081}
+}
+@book{ciarlet2002finite,
+  title={The finite element method for elliptic problems},
+  author={Ciarlet, Philippe G},
+  year={2002},
+  publisher={SIAM}
+}
+@book{reddy2019introduction,
+  title={Introduction to the finite element method},
+  author={Reddy, Junuthula Narasimha},
+  year={2019},
+  publisher={McGraw-Hill Education}
+}
+@article{cahn1958free,
+  title={Free energy of a nonuniform system. I. Interfacial free energy},
+  author={Cahn, John W and Hilliard, John E},
+  journal={The Journal of chemical physics},
+  volume={28},
+  number={2},
+  pages={258--267},
+  year={1958},
+  publisher={American Institute of Physics},
+  doi={10.1063/1.1744102}
 }
