### 2022-10-20 #4252 @vijaysm
While the statement of need section reads much better now after new changes, I still think you should present some examples of usage and capabilities in the paper. It is quite incomplete without it IMO. It reads more like a proposal at the moment. @yangbai90

I also made more comments in the open issues in the repo. Please address them where appropriate.
