+++ src/ElmtSystem/MieheFractureElement.cpp
@@ -8,18 +8,19 @@
 //****************************************************************
 //++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 //+++ Author : Yang Bai
-//+++ Date   : 2021.12.25
-//+++ Purpose: implement the residual and jacobian for allen-cahn 
-//+++          type phase-field fracture model
+//+++ Date   : 2022.10.25
+//+++ Purpose: implement the residual and jacobian for phase-field fracture 
+//             model based on Prof. Miehe's CMAME paper
 //+++          1) dD/dt=-L*(delta f/delta D)
 //+++          2) div(\Sigma)=0
-//+++ Ref    : A continuum phase field model for fracture, by Prof. M\"uller
-//+++ DOI    : https://doi.org/10.1016/j.engfracmech.2010.08.009
+//+++ Ref    : A phase field model for rate-independent crack propagation: 
+//             Robust algorithmic implementation based on operator splits
+//+++ DOI    : https://doi.org/10.1016/j.cma.2010.04.011
 //++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-#include "ElmtSystem/AllenCahnFractureElement.h"
+#include "ElmtSystem/MieheFractureElement.h"
 
-void AllenCahnFractureElement::computeAll(const FECalcType &calctype,const LocalElmtInfo &elmtinfo,const double (&ctan)[3],
+void MieheFractureElement::computeAll(const FECalcType &calctype,const LocalElmtInfo &elmtinfo,const double (&ctan)[3],
             const LocalElmtSolution &soln,const LocalShapeFun &shp,
             const MaterialsContainer &mate_old,const MaterialsContainer &mate,
             MatrixXd &localK,VectorXd &localR) {
@@ -30,12 +31,12 @@ void AllenCahnFractureElement::computeAll(const FECalcType &calctype,const Local
         computeJacobian(elmtinfo,ctan,soln,shp,mate_old,mate,localK);
     }
     else{
-        MessagePrinter::printErrorTxt("unsupported calculation type in AllenCahnFractureElmt, please check your related code");
+        MessagePrinter::printErrorTxt("unsupported calculation type in MieheFractureElement, please check your related code");
         MessagePrinter::exitAsFem();
     }
 }
 //***************************************************************************
-void AllenCahnFractureElement::computeResidual(const LocalElmtInfo &elmtinfo,
+void MieheFractureElement::computeResidual(const LocalElmtInfo &elmtinfo,
                                  const LocalElmtSolution &soln,
                                  const LocalShapeFun &shp,
                                  const MaterialsContainer &mate_old,
@@ -47,20 +48,20 @@ void AllenCahnFractureElement::computeResidual(const LocalElmtInfo &elmtinfo,
     if(mate_old.getScalarMaterialsNum()) {}
 
     if(elmtinfo.m_dim<2){
-        MessagePrinter::printErrorTxt("AllenCahnFractureElmt works only for 2d and 3d case");
+        MessagePrinter::printErrorTxt("MieheFractureElement works only for 2d and 3d case");
         MessagePrinter::exitAsFem();
     }
-    double L=mate.ScalarMaterial("L");
[…]
