+++ include/MateSystem/MateType.h
@@ -45,6 +45,7 @@ enum class MateType{
     CAHNHILLIARDMECHANICSMATE,
     SMALLSTRAINDIFFUSIONMATE,
     LINEARELASTICFRACMATE,
+    MIEHEFRACTUREMATE,
     // For user-defined-materials(UMAT)
     USER1MATE,
     USER2MATE,
