+++ src/MateSystem/RunBulkMateLibs.cpp
@@ -71,6 +71,12 @@ void BulkMateSystem::runBulkMateLibs(const MateType &t_matetype,const nlohmann::
     case MateType::LINEARELASTICFRACMATE:
         LinearElasticFractureMaterial::computeMaterialProperties(t_params,t_elmtinfo,t_elmtsoln,m_materialcontainer_old,m_materialcontainer);
         break;
+    case MateType::MIEHEFRACTUREMATE:
+        MieheFractureMaterial::computeMaterialProperties(t_params,t_elmtinfo,t_elmtsoln,m_materialcontainer_old,m_materialcontainer);
+        break;
+    case MateType::CAHNHILLIARDMECHANICSMATE:
+        CahnHilliardMechanicsMaterial::computeMaterialProperties(t_params,t_elmtinfo,t_elmtsoln,m_materialcontainer_old,m_materialcontainer);
+        break;
     default:
         MessagePrinter::printErrorTxt("Unsupported material type in runBulkMateLibs, please check your code");
         MessagePrinter::exitAsFem();

[…]
