### 2022-10-26 3d7138b
Remove the abuse of std namespace according to @bbsy789 's comments. Remove dummy include headers.
