+++ include/FE/ShapeFunUserBase.h
@@ -23,7 +23,9 @@
 #include "Utils/MessagePrinter.h"
 #include "MathUtils/Vector3d.h"
 
-using namespace std;
+using std::vector;
+using std::abs;
+using std::sqrt;
 
 /**
  * The base class for user-defined shape function
