+++ src/Mesh/Msh2FileImporter.cpp
@@ -20,7 +20,7 @@ int Msh2FileImporter::getMaxMeshDim(const string &filename)const{
     ifstream in;
     string str;
     int maxdim;
-    in.open(filename,ios::in);
+    in.open(filename.c_str(),ios::in);
     if(!in.is_open()){
         MessagePrinter::printErrorTxt("can\'t read the .msh file("+filename+"),please make sure file name is correct"
                                       " or you have the access permission");
@@ -73,7 +73,7 @@ bool Msh2FileImporter::importMeshFile(const string &filename,MeshData &meshdata)
     vector<vector<int>> ElmtConn;
 
     ifstream in;
-    in.open(filename,ios::in);
+    in.open(filename.c_str(),ios::in);
     if(!in.is_open()){
         MessagePrinter::printErrorTxt("can\'t read the .msh file("+filename+"),please make sure file name is correct"
                                       " or you have the access permission");
