+++ include/Postprocess/PostprocessorType.h
@@ -88,15 +88,15 @@ enum class PostprocessorType{
     USER4NODALPPS,
     USER5NODALPPS,
     //
-    USER1SIDEPPS,
-    USER2SIDEPPS,
-    USER3SIDEPPS,
-    USER4SIDEPPS,
-    USER5SIDEPPS,
+    USER1SIDEINTEGRALPPS,
+    USER2SIDEINTEGRALPPS,
+    USER3SIDEINTEGRALPPS,
+    USER4SIDEINTEGRALPPS,
+    USER5SIDEINTEGRALPPS,
     //
-    USER1VOLUMEPPS,
-    USER2VOLUMEPPS,
-    USER3VOLUMEPPS,
-    USER4VOLUMEPPS,
-    USER5VOLUMEPPS
+    USER1VOLUMEINTEGRALPPS,
+    USER2VOLUMEINTEGRALPPS,
+    USER3VOLUMEINTEGRALPPS,
+    USER4VOLUMEINTEGRALPPS,
+    USER5VOLUMEINTEGRALPPS
 };

[…]
