### 2022-10-26 5427aba
enable compiler flags output
