### 2022-10-26 7aa355d
increase precision for pps value & small changes to side integral
