### 2022-10-27 8a69731
enable local elmtinfo in volume pps
