+++ src/Postprocess/VolumeIntegralRank4MatePostprocessor.cpp
@@ -18,10 +18,11 @@
 double VolumeIntegralRank4MatePostprocessor::computeVolumeIntegralValue(const int &dofid,
                                             const int &nodeid,
                                             const nlohmann::json &parameters,
+                                            const LocalElmtInfo &elmtinfo,
                                             const LocalShapeFun &shp,
                                             SolutionSystem &soln,
                                             ProjectionSystem &projsystem){
-    if(dofid||soln.getDofsNum()||projsystem.getNodesNum()) {}
+    if(dofid||soln.getDofsNum()||projsystem.getNodesNum()||elmtinfo.m_dim) {}
     
     if(!JsonUtils::hasOnlyGivenValues(parameters,vector<string>{"nodeid","rank4mate","i-index","j-index","k-index","l-index"})){
         MessagePrinter::printErrorTxt("Unsupported options in parameters of the VolumeIntegralRank4MatePostprocessor, "
