+++ src/Postprocess/VolumePostprocessor.cpp
@@ -17,13 +17,14 @@
 double VolumePostprocessor::computeVolumeIntegralValue(const int &dofid,
                                             const int &nodeid,
                                             const nlohmann::json &t_parameters,
+                                            const LocalElmtInfo &elmtinfo,
                                             const LocalShapeFun &t_shp,
                                             SolutionSystem &t_soln,
                                             ProjectionSystem &t_projsystem){
     if(dofid||nodeid||t_parameters.size()||t_shp.m_test||t_soln.getDofsNum()||t_projsystem.getNodesNum()) {}
     
     
-    m_ppsvalue=1.0;
+    m_ppsvalue=1.0/elmtinfo.m_nodesnum;
     
     return m_ppsvalue;
 }

[…]
