+++ src/Postprocess/ExecutePostprocess.cpp
@@ -69,6 +69,7 @@ void Postprocessor::executePostprocess(const Mesh &t_mesh,
         case PostprocessorType::VOLUMEINTEGRATEVECTORMATERIALVALUE:
         case PostprocessorType::VOLUMEINTEGRATERANK2MATERIALVALUE:
         case PostprocessorType::VOLUMEINTEGRATERANK4MATERIALVALUE:
+        case PostprocessorType::USER1VOLUMEINTEGRALPPS:
         {
             m_pps_values[i-1]=executeVolumeIntegralPostprocess(pps_type,dofid,getIthPPSBlock(i).m_domainnamelist,getIthPPSBlock(i).m_parameters,t_mesh,t_dofhandler,t_fe,t_solution,t_projsystem);
             break;
