### 2022-10-27 226315d
add boolean material support
