+++ include/MateSystem/MaterialsName.h
@@ -27,7 +27,8 @@ using std::map;
 using std::string;
 using std::vector;
 //*******************************************
-//*** for type redfine, in order to use its short name
+//*** for type redefine, in order to use its short name
+typedef map<string,bool>        BooleanMateType;/**< for boolean materials, this material cant be exported */
 typedef map<string,double>      ScalarMateType;/**< for scalar materials */
 typedef map<string,Vector3d>    VectorMateType;/**< for vector materials */
 typedef map<string,Rank2Tensor> Rank2MateType;/**< for rank-2 materials */
