+++ src/MateSystem/SmallStrainDiffusionMaterial.cpp
@@ -37,6 +37,13 @@ void SmallStrainDiffusionMaterial::computeMaterialProperties(const nlohmann::jso
     //**************************************************************
     if(mateold.getScalarMaterialsNum()){}
 
+    if(JsonUtils::hasValue(inputparams,"finite-strain")){
+        mate.BooleanMaterial("finite-strain")=JsonUtils::getBoolean(inputparams,"finite-strain");
+    }
+    else{
+        mate.BooleanMaterial("finite-strain")=false;// use small strain deformation as the default option
+    }
+
     m_Omega=JsonUtils::getValue(inputparams,"Omega");
     m_cref=JsonUtils::getValue(inputparams,"cref");
 
