+++ include/MateSystem/MateType.h
@@ -42,7 +42,7 @@ enum class MateType{
     // for wave material
     WAVEMATE,
     // For coupled material
-    CAHNHILLIARDMECHANICSMATE,
+    SMALLSTRAINCAHNHILLIARDMATE,
     SMALLSTRAINDIFFUSIONMATE,
     LINEARELASTICFRACMATE,
     MIEHEFRACTUREMATE,
