### 2022-10-27 5ddd417
update Autotest script
