+++ .github/workflows/main.yml
@@ -65,27 +65,27 @@ jobs:
         run: |
           cd temp
           ls -l
-          curl -L -O https://download.open-mpi.org/release/open-mpi/v4.1/openmpi-4.1.0.tar.gz
+          curl -L -O https://download.open-mpi.org/release/open-mpi/v4.1/openmpi-4.1.4.tar.gz
           echo "******************************************************"
-          echo "*** Openmpi 4.1.0 is download !"
+          echo "*** Openmpi 4.1.4 is download !"
           echo "******************************************************"
           ls -l
       - name: compile and install openmpi
         run: |
           echo "******************************************************"
-          echo "*** Start to compile openmpi-4.1.0"
+          echo "*** Start to compile openmpi-4.1.4"
           echo "******************************************************"
           cd temp
-          tar -xf openmpi-4.1.0.tar.gz
-          cd openmpi-4.1.0
+          tar -xf openmpi-4.1.4.tar.gz
+          cd openmpi-4.1.4
           ./configure --prefix=/home/runner/work/AsFem/AsFem/AsFemLibs/openmpi
           make -j8 > mymake.log
           echo "******************************************************"
-          echo "*** Make -j8 is done for openmpi-4.1.0"
+          echo "*** Make -j8 is done for openmpi-4.1.4"
           echo "******************************************************"
           make install > mymakeinstall.log
           echo "******************************************************"
-          echo "*** Make install is done for openmpi-4.1.0"
+          echo "*** Make install is done for openmpi-4.1.4"
           echo "******************************************************"
       - name: check openmpi version
         run: |
@@ -94,6 +94,7 @@ jobs:
           echo "*** Now we check the version of mpicc and mpicxx"
           ./mpicc --version
           ./mpicxx --version
+          ./mpirun --version
           echo "******************************************************"
       - name: downlaod PETSc
         run: |
@@ -101,11 +102,11 @@ jobs:
           echo "*** download petsc"
           echo "******************************************************"
           cd /home/runner/work/AsFem/AsFem/temp
-          curl -L -O  https://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.14.3.tar.gz
-          tar -xf petsc-lite-3.14.3.tar.gz
+          curl -L -O  https://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.17.5.tar.gz
+          tar -xf petsc-lite-3.17.5.tar.gz
       - name: compile and install PETSc
         run: |
-          cd /home/runner/work/AsFem/AsFem/temp/petsc-3.14.3
+          cd /home/runner/work/AsFem/AsFem/temp/petsc-3.17.5
           echo "******************************************************"
           echo "*** Config petsc ..."
           echo "******************************************************"
@@ -116,12 +117,8 @@ jobs:
           --with-pic=1 \
           --with-mpi-dir=/home/runner/work/AsFem/AsFem/AsFemLibs/openmpi \
           --with-shared-libraries=1 \
-          --with-cxx-dialect=C++14 \
           --with-fortran-bindings=0 \
           --with-sowing=0 \
-          --download-scalapack=1 \
-          --download-superlu_dist=1 \
-          --download-mumps=1 \
           COPTFLAGS='-fPIC -O3 ' \
           CXXOPTFLAGS='-fPIC -O3 ' \
           FOPTFLAGS='-fPIC -O3 ' \
@@ -129,11 +126,11 @@ jobs:
[…]
