### 2022-10-27 d36fe73
update clean script
