+++ .github/workflows/main.yml
@@ -67,7 +67,7 @@ jobs:
           ls -l
           curl -L -O https://download.open-mpi.org/release/open-mpi/v4.1/openmpi-4.1.4.tar.gz
           echo "******************************************************"
-          echo "*** Openmpi 4.1.4 is download !"
+          echo "*** Openmpi 4.1.4 has been downloaded !"
           echo "******************************************************"
           ls -l
       - name: compile and install openmpi
@@ -79,11 +79,11 @@ jobs:
           tar -xf openmpi-4.1.4.tar.gz
           cd openmpi-4.1.4
           ./configure --prefix=/home/runner/work/AsFem/AsFem/AsFemLibs/openmpi
-          make -j8 > mymake.log
+          make -j8
           echo "******************************************************"
           echo "*** Make -j8 is done for openmpi-4.1.4"
           echo "******************************************************"
-          make install > mymakeinstall.log
+          make install
           echo "******************************************************"
           echo "*** Make install is done for openmpi-4.1.4"
           echo "******************************************************"
@@ -102,11 +102,11 @@ jobs:
           echo "*** download petsc"
           echo "******************************************************"
           cd /home/runner/work/AsFem/AsFem/temp
-          curl -L -O  https://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.17.5.tar.gz
-          tar -xf petsc-lite-3.17.5.tar.gz
+          curl -L -O  https://ftp.mcs.anl.gov/pub/petsc/release-snapshots/petsc-lite-3.18.1.tar.gz
+          tar -xf petsc-lite-3.18.1.tar.gz
       - name: compile and install PETSc
         run: |
-          cd /home/runner/work/AsFem/AsFem/temp/petsc-3.17.5
+          cd /home/runner/work/AsFem/AsFem/temp/petsc-3.18.1
           echo "******************************************************"
           echo "*** Config petsc ..."
           echo "******************************************************"
@@ -119,6 +119,7 @@ jobs:
           --with-shared-libraries=1 \
           --with-fortran-bindings=0 \
           --with-sowing=0 \
+          --download-fblaslapack=1 \
           COPTFLAGS='-fPIC -O3 ' \
           CXXOPTFLAGS='-fPIC -O3 ' \
           FOPTFLAGS='-fPIC -O3 ' \
@@ -126,11 +127,11 @@ jobs:
           echo "******************************************************"
           echo "*** Make petsc ..."
           echo "******************************************************"
-          make PETSC_DIR=/home/runner/work/AsFem/AsFem/temp/petsc-3.17.5 PETSC_ARCH=arch-linux-c-opt all > mymake.log
+          make PETSC_DIR=/home/runner/work/AsFem/AsFem/temp/petsc-3.18.1 PETSC_ARCH=arch-linux-c-opt all
           echo "******************************************************"
           echo "*** Install petsc ..."
           echo "******************************************************"
-          make PETSC_DIR=/home/runner/work/AsFem/AsFem/temp/petsc-3.17.5 PETSC_ARCH=arch-linux-c-opt install > mymakeinstall.log
+          make PETSC_DIR=/home/runner/work/AsFem/AsFem/temp/petsc-3.18.1 PETSC_ARCH=arch-linux-c-opt install
       - name: compile AsFem
         run: |
           cd /home/runner/work/AsFem/AsFem
@@ -139,6 +140,10 @@ jobs:
           echo "*********************************************************************"
           export MPI_DIR=/home/runner/work/AsFem/AsFem/AsFemLibs/openmpi
           export PATH=$PATH:$MPI_DIR/bin
[…]
