### 2022-10-27 1a44390
fix path error in AutoTest script
