+++ include/MateSystem/MateType.h
@@ -45,6 +45,7 @@ enum class MateType{
     SMALLSTRAINCAHNHILLIARDMATE,
     SMALLSTRAINDIFFUSIONMATE,
     LINEARELASTICFRACMATE,
+    NEOHOOKEANPFFRACTUREMATE,
     MIEHEFRACTUREMATE,
     // For user-defined-materials(UMAT)
     USER1MATE,
