+++ include/MateSystem/NeoHookeanPFFractureMaterial.h
@@ -1,0 +1,147 @@
+//****************************************************************
+//* This file is part of the AsFem framework
+//* A Simple Finite Element Method program (AsFem)
+//* All rights reserved, Yang Bai/M3 Group@CopyRight 2020-present
+//* https://github.com/M3Group/AsFem
+//* Licensed under GNU GPLv3, please see LICENSE for details
+//* https://www.gnu.org/licenses/gpl-3.0.en.html
+//****************************************************************
+//++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+//+++ Author : Yang Bai
+//+++ Date   : 2022.10.28
+//+++ Purpose: Calculate the material properties defined in neohookean
+//+++          type finite strain phase-field fracture model
+//++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+#pragma once
+
+#include "MateSystem/BulkMaterialBase.h"
+#include "MateSystem/FreeEnergyMaterialBase.h"
+#include "MateSystem/ElasticMaterialBase.h"
+
+/**
+ * This class calculate the material properties for the allen-cahn fracture element
+ */
+class NeoHookeanPFFractureMaterial:public BulkMaterialBase,
+                                   public FreeEnergyMaterialBase,
+                                   public ElasticMaterialBase{
+public:
+    NeoHookeanPFFractureMaterial();
+    ~NeoHookeanPFFractureMaterial();
+protected:
+    /**
+     * Initial the preset material properties, if you don't need the history information of some materials, 
+     * then you can avoid calling this function
+     * @param t_inputparams the input material parameters read from the json file
+     * @param t_elmtinfo the data structure for the local element information
+     * @param t_elmtsoln the solutions, i.e., U and V of the local element
+     * @param Mate the materials (container) to be initialized
+     */
+    virtual void initMaterialProperties(const nlohmann::json &t_inputparams,
+                                        const LocalElmtInfo &t_elmtinfo,
+                                        const LocalElmtSolution &t_elmtsoln,
+                                        MaterialsContainer &t_mate) override;
+
+    /**
+     * Compute the material property accroding to your model
+     * @param t_inputparams the input material parameters read from the input file
+     * @param t_elmtinfo the data structure for the local element information
+     * @param t_elmtsoln the solutions, i.e., U and V of the local element
+     * @param t_mateold the materials from previous step
+     * @param t_mate the materials to be calculated
+     */
+    virtual void computeMaterialProperties(const nlohmann::json &t_inputparams,
+                                           const LocalElmtInfo &t_elmtinfo,
+                                           const LocalElmtSolution &t_elmtsoln,
+                                           const MaterialsContainer &t_mateold,
+                                           MaterialsContainer &t_mate) override;
+
+private:
+    /**
+     * compute the elastic strain tensor
+     * @param dim the dimension of current analysis
+     * @param gradU the gradient of displacement
+     * @param strain the output elastic strain tensor
+     */
+    virtual void computeStrain(const int &dim,const Rank2Tensor &gradU,Rank2Tensor &strain) override;
+    /**
+     * compute the stress and jacobian tensor
+     * @param t_params the json parameters defined in the input file
+     * @param dim the dimension of current analysis
+     * @param strain the strain tensor
+     * @param stress the output stress tensor
+     * @param jacobian the output jacobian tensor
+     */
+    virtual void computeStressAndJacobian(const nlohmann::json &t_params,
+                                          const int &dim,
[…]
