+++ src/MateSystem/NeoHookeanPFFractureMaterial.cpp
@@ -13,23 +13,23 @@
 //+++          element. In this code, we can define:
 //++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-#include "MateSystem/LinearElasticFractureMaterial.h"
+#include "MateSystem/NeoHookeanPFFractureMaterial.h"
 #include "MathUtils/MathFuns.h"
 
-LinearElasticFractureMaterial::LinearElasticFractureMaterial(){
+NeoHookeanPFFractureMaterial::NeoHookeanPFFractureMaterial(){
     m_args.resize(11);
     m_F.resize(11);
     m_dFdargs.resize(11);
     m_d2Fdargs2.resize(11,11);
 }
-LinearElasticFractureMaterial::~LinearElasticFractureMaterial(){
+NeoHookeanPFFractureMaterial::~NeoHookeanPFFractureMaterial(){
     m_args.resize(11);
     m_F.resize(11);
     m_dFdargs.resize(11);
     m_d2Fdargs2.resize(11,11);
 }
 //******************************************************
-void LinearElasticFractureMaterial::initMaterialProperties(const nlohmann::json &inputparams,
+void NeoHookeanPFFractureMaterial::initMaterialProperties(const nlohmann::json &inputparams,
                                         const LocalElmtInfo &elmtinfo,
                                         const LocalElmtSolution &elmtsoln,
                                         MaterialsContainer &mate){
@@ -40,7 +40,7 @@ void LinearElasticFractureMaterial::initMaterialProperties(const nlohmann::json 
     mate.ScalarMaterial("H")=0.0;
 }
 //********************************************************************
-void LinearElasticFractureMaterial::computeMaterialProperties(const nlohmann::json &inputparams,
+void NeoHookeanPFFractureMaterial::computeMaterialProperties(const nlohmann::json &inputparams,
                                            const LocalElmtInfo &elmtinfo,
                                            const LocalElmtSolution &elmtsoln,
                                            const MaterialsContainer &mateold,
@@ -52,13 +52,13 @@ void LinearElasticFractureMaterial::computeMaterialProperties(const nlohmann::js
 
     if(JsonUtils::hasValue(inputparams,"finite-strain")){
         mate.BooleanMaterial("finite-strain")=JsonUtils::getBoolean(inputparams,"finite-strain");
-        if(mate.BooleanMaterial("finite-strain")){
-            MessagePrinter::printErrorTxt("Linear elastic fracture material works only in small strain case. Please check your input file and set 'finite-strain'=false");
+        if(!mate.BooleanMaterial("finite-strain")){
+            MessagePrinter::printErrorTxt("For neohookean fracture material, you must enable finite-strain option. Please check your input file");
             MessagePrinter::exitAsFem();
         }
     }
     else{
-        mate.BooleanMaterial("finite-strain")=false;// use small strain deformation as the default option
+        mate.BooleanMaterial("finite-strain")=true;// use finite strain deformation as the default option
     }
 
     mate.ScalarMaterial("L")=JsonUtils::getValue(inputparams,"L");
[…]
