### 2022-10-28 6532ce7
optimize param print & explicitly define finite-strain option in json
