### 2022-10-28 #4252 @Kevin-Mattheus-Moerman
@yangbai90 I understand, but it is very unfortunate. We do expect authors to be committed to improving and maintaining the software, especially during the review process. I understand you decision but would ask you to not submit works to JOSS in the future unless the authors have the bandwidth to maintain the software and respond to review comments.
